Обновлены файлы в MapInfo
Поля заменены на свойства, которые ограничивают вводимое значение.
Доступ к списку Bookmarks можно получить только как ReadOnlyCollection.

+++ Assets/OsuEditor/Settings/Difficulty/STSlider.cs
@@ -26,7 +26,7 @@ namespace Assets.OsuEditor.Settings.Difficulty
 
         void SaveValue()
         {
-            Global.Map.Difficulty.SliderTickRate = thisSlider.value;
+            Global.Map.Difficulty.SliderTickRate = (int)thisSlider.value;
             STText.text = STText.GetComponent<LangWriter>().GetText() + " - " + thisSlider.value;
         }
     }
